backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Cover letter generator API"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "authlib>=1.6.8",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.129.2",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.15.2",
    "types-beautifulsoup4>=4.12.0.20250516",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
mypy_path = "src"
packages = ["src"]
